backend/requirements.txt
google-api-python-client==2.123.0
google-auth==2.29.0
flask==2.3.3
openai==1.74.0
requests==2.31.0
elevenlabs==1.56.0
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
celery[redis]
redis
pydantic[email]
python-multipart
razorpay
msgpack[accelerated]
